requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-filter==0.13.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
Markdown==2.6.6
markdown2==2.3.1
Pillow==6.0.0
PyJWT==1.4.0
